+++ src/main/java/de/samply/reporter/exporter/RequestResponseEntity.java
@@ -1,0 +1,7 @@
+package de.samply.reporter.exporter;
+
+public record RequestResponseEntity(
+    String responseUrl
+) {
+
+}