+++ src/main/java/de/samply/reporter/utils/FileUtils.java
@@ -1,0 +1,14 @@
+package de.samply.reporter.utils;
+
+import de.samply.reporter.app.ReporterConst;
+import org.apache.commons.lang3.RandomStringUtils;
+
+public class FileUtils {
+
+  public static String fetchRandomFilename(String fileFormatExtension) {
+    return RandomStringUtils.random(ReporterConst.RANDOM_FILENAME_SIZE, true, false) + "."
+        + fileFormatExtension;
+  }
+
+
+}